- Add Transaction entities and repositories. - Add Transaction Maintainer. - Add delete product conditions.

+++ app-dashboard/src/main/java/cl/example/dashboard/services/ProductService.java
@@ -2,7 +2,9 @@ package cl.example.dashboard.services;
 
 import cl.example.entities.domain.entities.ClientEntity;
 import cl.example.entities.domain.entities.ProductEntity;
+import cl.example.entities.domain.entities.TransactionProductEntity;
 import cl.example.entities.domain.repositories.ProductRepository;
+import cl.example.entities.domain.repositories.TransactionProductRepository;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import cl.example.entities.domain.entities.vo.ProductVO;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -22,6 +25,9 @@ public class ProductService {
 
     @Autowired
     private EntityManager entityManager;
+
+    @Autowired
+    private TransactionProductRepository transactionProductRepository;
 
     public ProductVO getProductById(int productId, int clientId) {
         ProductEntity entity = productRepository.findByIdAndClient_Id(productId, clientId);
@@ -59,6 +65,11 @@ public class ProductService {
     }
 
     public void deleteProduct(int productId, int clientId) {
+        List<TransactionProductEntity> tps = transactionProductRepository.findById_Product_Id(productId);
+        if (!tps.isEmpty()) {
+            throw new RuntimeException("Unable to delete product if it already has transactions.");
+        }
+
         productRepository.deleteByIdAndClient_Id(productId, clientId);
     }
